feat: create a mockData file for frontend components
